backend/requirements.txt
requests
djangorestframework
django-cors-headers
python-dotenv